+++ src/status/index.ts
@@ -3,4 +3,5 @@ export { StatusBarManager } from './statusBarManager';
 export { BaseStatusBarItem, BaseStatusBarItemConfig, StatusBarItemConfig, CachedStatusData } from './baseStatusBarItem';
 export { ProviderStatusBarItem } from './providerStatusBarItem';
 export { AntigravityStatusBar } from './antigravityStatusBar';
+export { MistralStatusBar } from './mistralStatusBar';
 export { registerCombinedQuotaCommand } from './combinedQuotaPopup';